[U] fix des trucs (kotlin me clc)

+++ CM_IMMOAPP/app/src/main/java/com/cm_immo_app/models/PropertySimple.kt
@@ -3,7 +3,7 @@ package com.cm_immo_app.models
 data class PropertySimple(
     val id: String,
     val name: String,
-    val completion: Int,
+    val completion: Float,
     val ImageUrl: String,
     val location: String
 )
